+++ lib/open_project/reporting/engine.rb
@@ -62,7 +62,7 @@ module OpenProject::Reporting
           author_url 'https://www.finn.de'
           version OpenProject::Reporting::VERSION
 
-          requires_openproject ">= 3.0.0pre23"
+          requires_openproject ">= 3.0.0pre30"
 
           view_actions = [:index, :show, :drill_down, :available_values, :display_report_list]
           edit_actions = [:create, :update, :rename, :delete]